+++ CppScripts/QtransportLayerAgentN.cpp
@@ -767,8 +767,8 @@ for (int iIterMessages=0;iIterMessages<NumQintupleComas;iIterMessages++){
 			char PayloadAux[NumBytesPayloadBuffer]={0};
 			strcpy(PayloadAux,strtok(NULL,";"));
 			this->QLLAnumReqQuBits=atoi(strtok(NULL,";"));// Copy this first to not lose strtok pointer
-			this->QLLAFineSynchAdjVal[0]=atoi(strtok(NULL,";"));// Copy this first to not lose strtok pointer
-			this->QLLAFineSynchAdjVal[1]=atoi(strtok(NULL,";"));// Copy this first to not lose strtok pointer
+			this->QLLAFineSynchAdjVal[0]=atof(strtok(NULL,";"));// Copy this first to not lose strtok pointer
+			this->QLLAFineSynchAdjVal[1]=atof(strtok(NULL,";"));// Copy this first to not lose strtok pointer
 			int numUnderScores=countUnderscores(PayloadAux);
 			for (int iIterQLLAIPaddr=0;iIterQLLAIPaddr<numUnderScores;iIterQLLAIPaddr++){
 				if(iIterQLLAIPaddr==0){strcpy(this->QLLAIPaddresses[iIterQLLAIPaddr],strtok(PayloadAux,"_"));}
